Changed the way focus-out-events are handled in gui logging.
Focus-out-events are tricky to log because they can be generated for
widgets that no longer exist (or are no longer in their containers).
The gui log files used to contain lines like:
   widget = findWidget(...)
   if widget:
      do something with widget
   del widget
This was ugly.  Now it can be done like this:
   widget = weakRef(findWidget(...))
   if widget():
      do something with widget()
weakRef is a function that returns a weak reference to its argument,
unless its argument is None, in which case it returns a function
that returns None.  It's no longer necessary to delete widget.  More
importantly, it's now possible to reuse the local variables, since
they're not deleted.  When recording, gtklogger keeps a weak key
dictionary of the widgets that have been used, and reuses them if
the referend still exists when they're needed again.