Cleaned up error handling wrt gui logging a bit.
The excepthook argument to gtklogger.replay() no longer seems
to be necessary, at least in the way oof2 uses it, so I commented
it out of guilogger.py.  It's still in gtklogger/replay.py because
it might be needed if anyone else ever uses gtklogger.  I updated
the README.

gtklogger.replay() now re-raises exceptions that it catches, after
aborting the log player.

Added a divide by zero test in the Help/Debug/Error menu, to see
how gtklogger would handle it.