Fixed a TODO about alignment of widgets within MatrixInput and MatrixBoolInput.
The GtkEntrys for setting Cijkl components (for example) expand
to fill their space when their window grows and therefore stay
centered in their columns.  The checkboxes for a MatrixBoolInput
(a matrix of checkboxes) do not expand, and weren't centered in
their columns.  Now the alignment is set in the derived class and
is correct in both cases.