Changed the way that graphics layers are ordered.
If a new layer is not replacing an existing layer, it is positioned
at the lowest position in the layer stack such that its sorting
order is lower than all the layers above it.  For example, a new
contour layer will be displayed below all the skeleton edge layers,
unless there's already a contour layer above the skeleton edges,
in which case the new layer is displayed above the old one.  This
makes new layers as visible as possible without hiding anything that's
not already hidden.

Added a paragraph about reordering layers to the Simple tutorial.

Fixed some gui tests that were broken by this change and ones that
were broken when I added a gtklogger widgetname to a pane in the
Analysis Page.