Fixed a bug in the Microstructure page, and rerecorded gui tests affected by it.
The state of the "meshable" check button was being set by the routine
that rebuilds the list of pixel groups, but in some cases it was
doing it before knowing which group was selected.  This could cause
the button to be set incorrectly.  For some reason the problem
only appeared when running oof2 in unthreaded mode.

Now the selection, if there is one, is passed in to the routine
that rebuilds the list.  The meshable button is set more efficiently
(ie, less often).  Checkpoints associated with the meshable button
therefore changed, and lots of gui tests had to be rerecorded.