Debugging new rubberband code on a new branch.
There had been a TODO about a possible race condition with
rubberbands, and I finally found a situation that triggers it.
Rubberbands in the Move Nodes toolbox don't appear when they
should because the first draw occurs before the rubberband is
constructed, and the next draw doesn't occur until after the
mouse is moved.  The solution from the TODO is to have the
mouse down callback return the rubberband