Changed the Rationalize skeleton modifier to eliminate a possible infinite loop.
The previous method put newly created elements on the list of elements
yet to be examined.  This caused an infinite loop in rare circumstances.
Now the new elements aren't examined, which means that a single pass
through the skeleton does less work than it did before.  To make
up for that, the whole process is repeated a set number of times,
given by an "iterations" argument.