Rubberbands defined in OOFCanvas don't try to draw if their layers are null.
Because gtklogger has to bypass a lot of the OOFCanvas machinery,
as explained by a comment in gfxwindowbase.py, rubberbands on
the OOFCanvas aren't initialized properly when replaying a gtklogger
log file. Having the rubberbands just check before they use a null
layer is the easiest hack to avoid the problem.