Modifications to make the compiler on Big Sur happy.
Lots of warnings about losing precision when assigning a long
(usually a std::vector::size_type) to an int have been eliminated,
mostly by replacing the int with a std::size_t. In the few cases
where we really wanted an int, I added an assert to make sure
that the value of the long wasn't too large.

I don't understand why *all* of the above changes were in the
common directory.

Also removed some unused code, rather than fix the int/long
problems in it.  Apparently at some time we were considering
allowing Microstructures to be resized.  There was a resize
method for PixelGroups, but not for Microstructures.