Various TODOs and other improvements to OOFCanvas.
The typemap for OOFCanvas::Coords passed in to C++ from Python now
just assumes that the incoming type can be indexed, and uses
coord[0] and coord[1] to extract x and y.  This is more
generic than assuming that the object has x and y attributes.  It
also means that a Coord passed out to Python as a tuple can be
passed back in, and doesn't require swigging OOFCanvas::Coord.

Added more constructors and ways of adding points to CanvasPolygon and
CanvasCurve.  They're more consistent now.  Constructors can take
vectors of Coords.  Single Coords and vectors of Coords can be added
to an existing polygon or curve.  In Python, a single point can be
added with addPoint(x,y), addCoord(coord), or
addPoints([coord, coord, ...]), where coord is an indexable object.

Removed the unused TMatrix class from utility.[Ch].

Added OOFCanvas::Coord::operator[](int).

mouseCallbackData is passed to the callback function by Canvas::doCallback.