Changed some OOFCanvas function calls because they were badly named and confusing to use.
The old way of setting a line width in pixel units was to call
CanvasShape::setLineWidth(w) and then convert it to pixels with
CanvasShape::setLineWidthInPixels().  When writing the documenation
I realized how dumb this is.  Now setLineWidth(w) sets the width
in physical units and setLineWidthInPixels(w) sets it in pixel
units.  The last one called is the one that takes effect.

Similar changes were made for setting dash lengths.