OOFCanvas changes for simplicity and consistency.
Since Color and Coord are swigged, there's no need to avoid using
them for arguments and return values in swigged functions.

In python, Canvas.clickedItems() converts its return value into
a python list, instead of using a specialized iterator to loop over
the C++ list.  It's not a time-critical function, and this way
is simpler.

Added a TODO about using pythonexportable in OOFCanvas.  The hard-coded
module names that it requires will be a problem if OOFCanvas is going
to be made available outside of OOF2.