Added some robustness to setup.py.
These changes don't affect oof2 at the moment, but fix problems that
arose when I tried to use oof2's setup.py to install a standalone
version of OOFCanvas.

find_swig_pkgs() was failing if there were swig files directly
in SRC, instead of in SRC/common.  This was because it assumed that
it could strip the first 6 characters off of a path.  The OOFCanvas
path was "./SRC", which didn't have six characters.  The new version
uses os.path.relpath() instead of just cropping the string.

clean_dependencies() was assuming that the build temp directory
already existed, which was only true because oofconfig.h had already
been built, and had constructed the directory.  Now clean_dependencies()
uses disutils.dir_util.mkpath() to ensure that the directory exists.