Changes to get OOF2 to work (sort of) when OOFCanvas is installed separately.
The canvas appears and can be drawn to, but there are issues.

swiglib.C is included in OOF2 *and* in OOFCanvas.  This makes OOF2
crash.  I think it's because SWIG_GetPtr caches values and the cache
is corrupted when there are two copies of it, or something like that.
In any case, this version runs only if OOFCanvas is rebuilt *without*
swiglib.C.  I'm not sure of the right way to fix this.  Perhaps swig
also needs to be in a standalone library.

OOF2 functions that return a swigged OOFCanvas object need new
typemaps.  When OOFCanvas was part of OOF2, a simple
   %extern "oofcanvas/oofcanvas.swg"
worked, but now that swig file is not accessible when building OOF2.

Because OOFCanvas is installed in my home directory, I added ~/include
and ~/lib to the compile and link search paths in setup.py.  This and
a few issues in the DIR.py files could be solved by having OOFCanvas
create an oofcanvas.pc file for pkg-config.