Changes to get OOF2 to work with the latest improved OOFCanvas.
Not everything is working, but the canvas can display images
and skeletons.

The DIR.py files now get flags and libs from "pkg-config oofcanvas"
and don't refer to cairo, pango, or ImageMagick unless OOF2 is using
them explicitly.  The hack that added some flags in setup.py has
been removed.

The minimum required OOFCanvas version number is in the setup.py
file, currently set to "0.0.0".

canvasimage.swg contains typemaps for exporting an OOFCanvas::CanvasImage
to Python.  OOF2 needs to do this, but can't access the swig typemaps
used in the OOFCanvas source code.