Fixed some ancient and possibly inconsequential swig typemap errors.
Two calls to SWIG_GetPtr passed the wrong string for the expected
object class, which made type checking fail.  But that was ok, sort
of, because SWIG_GetPtr's return value was being ignored.  The errors
are also in OOF3D, which means they are very old.

I changed SWIG_GetPtr so that it throws an exception instead of
returning an error code, but only when compiled in debug mode.