Made openmp optional even if compiled with -fopenmp.
ImageMagick on Ubuntu now builds with -fopenmp, and we use
ImageMagick's c++ flags (via pkg-config), so autogroup was
using openmp whether we wanted it to or not.  There apparently are bugs
in autogroup when build with openmp, and I don't want to fix them now,
so I just put all of the omp pragmas inside #ifdef HAVE_OPENMP blocks.
HAVE_OPENMP is set, or not, in oofconfig.h and determined by the
--enable-openmp option in setup.py.