Fixed a regression test interaction bug.
The iterations argument added to the Rationalize Skeleton modifier
had been added to skeleton_basic_test but not skeleton_periodic_test,
so when skeleton_periodic_test was run by itself it used a different
number of iterations than when it was run after skeleton_basic_test.

I don't know why this was only caught now, and not when Rationalize
was modified.  Perhaps the tests couldn't be run until the gtk3
conversion was complete?