Changed tests so that the common code is all in regression.py.
This copies changes that were made in oof3d long ago.  Each test
file just needs to contain a list named 'test_set' containing
the unittest objects.

The reason for doing this now is that changes need to be made so
that the tests can be installed in some standard place and run
easily even if oof2 is installed by macports.  The paths to the
test files are not obvious in that case, so we can't tell users
to simply cd to the directory and run regression.py.