OOF2 tests can now be run from any directory with the scripts oof2-test and oof2-guitest.
The scripts are installed in bin along with oof2.  They just wrap
regression.py and guitests.py.

Also added a --list option to both regression.py and guitests.py, since
otherwise it's hard to know what the names of the tests are.

Also added --help and improved the output when an illegal option is provided.