More conversion from char[] to std::string, to prevent errors when file paths are too long.
There are many places where the conversion is incomplete, in that
instead of passing a std::string to a function, its c_str() is
passed instead.  This saves having to convert the functions, although
it's inelegant.  Some function args needed to be converted from
char* to const char*, though.

Some parts of main.cxx that we don't use were deleted or commented
out so that they didn't have to be converted, or just to simplify
the code to make it easier to work with.  This includes the portions
inside #ifdef MACSWIG and code for generating documentation.  MACSWIG
seems to have been there to support pre-OSX Macs, with ':' used
instead of '/' in paths.

Added -std=c++11 to the SWIG build options.