Deleted more unused code.
Most of the machinery for generating documentation has
been deleted.  Some of it is mentioned in parser.y, and I
don't want to edit that file without knowing what I'm doing,
so the CommentHandler class and comment_handler instance
have been retained.

Deleted the swig command line options related to documentation,
and removed --dnone from the swig calling sequence in oof2setuputils.py.

Delete the options for checking files into and out of the swig
library, which we're not using.