Un-reverted (actually just typed) a fix that shouldn't have been reverted.
Previous "git revert"s were to undo the pruning that I did to
the parts of swig that we don't use.  It turns out that they can't
be deleted without also modifying parser.y, and since I don't
understand parser.y, I don't want to that.

I'm not sure why the pruning appeared to work.  Perhaps "oof2-build
dist --swig" isn't working properly and an old libswig.a was being
used.  In any case, libswig.a refers to symbols in the swig Doc*
classes, so they have to exist even if they're not used.

Fixing this neatly is more effort than it's worth, since we hope
to stop using this version of swig.