Changed make_dist so that releases can be created from branches.
The old script forbid releases from branches, for a poor reason
having to do with merge conflicts and version numbers.  What should
be forbidden is doing development on and merging the automatically
created branch that's made for each release.  The root of that
branch doesn't have to be master.

I've edited the comments and re-enabled branch releases.