Better time zone handling for pdf files, I hope.
The pdf files created by Cairo (via OOFCanvas) include a time stamp,
which must be ignored by the test routines when comparing files,
because the test files aren't created at the same time as the
reference files.  pdf_compare in file_utils.py skips the time stamp.
Unfortunately, the pdf files also contain their lengths, and if the
format of the time stamp changes, then the length changes.  It's
harder to ignore the length because it's just a number.  This made
all the pdf file tests fail on nanoHUB, where the format of the
time stamp didn't include an explicit time zone.

Setting the time zone to "Etc/UTC" (apparently) forces the time
stamp into a time-zoneless format.  regression.py now sets the TZ
environment variable to "Etc/UTC" before starting oof2.  All reference
pdfs have been regenerated with the same TZ, and comments have been
added to the test scripts so that when those files have to be
regenerated and the tests start failing, we'll know to set TZ before
regenerating them again.