More progress on converting to cmake.
This is still the python2 version!

This version doesn't do swig, since there's probably no point in
figuring out how to use swig1.1 with cmake, if we're goign to try
to use a more modern swig or swig replacement.

This version will build, on the mac at least.  But because it doesn't
do swig, you first have to build the old distutils version to create
the swig output files.

The regression tests work.
The GUI tests work except for the error handling ones.