Working on converting typemaps to Python 3 and Swig 4.
Some new swig features are handy, like macros.
I defined a macro, MAKE_LISTVEC_TYPEMAPS, that creates typemaps
for converting a python list of swigged objects to a c++ vector
of the objects, and vice versa.  The C++ -> python typemap has
two versions, one which transfers ownership of the objects to python
and one which doesn't.