More progress on typemaps.
Added a FETCH_CLASS swig macro for use in those typemaps where
we need to fetch a python class object from a module.

We can't use the code that swig generates for us when it's given
the -threads option, so I copied it into pythonlock.h and modified
it so that threading can be turned on and off at run time.  This
will replace the old acquirePyLock and releasePyLock calls that
we used before.  The big difference is that the state is saved in
a class and restored by the destructor, making it easier to use if
exceptions might occur.