Progress on updating the swig files.  The process is not complete.
Deleted engine/typemaps.swg because %including it was requiring
many irrelevant header files to be required when compiling the
swig-generated C++ files.  The typemaps that used to be in it
are now in the swig files where their classes are defined.  Other
files that need the typemaps can just %import (not %include) the
relevant files.

common/typemaps.swg doesn't require the swig-generated files to
include a lot of C++ headers, so it can stay.