Moved SRC/engine/property to SRC/engine/properties.
Because we couldn't have a "property" subdirectory and a "property.py"
module in engine, we had used "properties.swg" instead of "property.swg".
This made properties.swg different from every other swig file, which
all have the same base name as their associated .C and .h files.
Now this is causing a problem, because cmake uses the %module line
in the .swg file to determine the name of the output file, and is
confused by the line "%module property" in properties.swg.

It would have made more sense anyway to put the Properties
in a "properties" subdirectory, and to call the swig file
"properties.swg".  So that's what I'm doing now.