New much simpler ScriptLoader class.
Either python3 doesn't have the problems that the python2
ScriptLoader was working around, or we were doing a lot of extra
work before, or maybe this new lighter version has problems that
haven't shown up yet.  In any case, the old version used some
library modules that don't exist in python3 so it had to be
rewritten.

The new version doesn't yet try to display errors and traces in
the GUI.