Fixed some %import and %pythoncode lines in swig files.
%imports in swig files generate imports in python files, which
can lead to import loops, and problems with undefined objects in
partially imported modules.  The solution in this case was to
remove an unnecessary import.

Moved a bunch of code out of femesh.spy and into meshiterator.spy,
because it didn't belong in femesh, and because of the new import
order was trying to modify a class that didn't yet exist.