Fixed indentation problems in subproblemcontext.py and matrixmethod.py.
Is python2 more lenient than python3 about indentation?  Some lines
were indented with a mix of tabs and spaces that was inconsistent
with the other lines in the file, and python3 barked at them.
(In subproblemcontext.py, the lines were the utils.memusage()
calls, which were inserted later than the rest, and possibly by
a different person (Adam C?) with a different editor.  I fixed
the indentation and commented out the lines.  I also commented
out the memusage line in evolve.py)

Also re-fixed an %import that I had broken on the previous commit.