Python3/Swig4 fixes to the MasterElement class.
The old version added python code to the MasterElement class
by hacking the swig-generated __init__.  That doesn't work now.
The code is added when each C++ MasterElement singleton is
exported to Python, which happens just once for each subclass.

Also fixed a sort(), which needed to switch to the python3 way of
doing things with a key function instaed of a __cmp__.