Various minor python3 fixes.
Use isinstance to check for the presence of a metaclass, not "if cls.__metaclass__ is ..."

The __bool__ method must return True or False, not 1 or 0.

There is no more types.InstanceType.

utils.PrintableClass was simplified. (Maybe it could have been in Python2).

There is no more lower() function in the string module. Use the
str.lower method.

fiddlenodesbase.py had mixed tab/space indentation.