Changed how python handles C++ exceptions.
In Python3, exceptions must be subclasses of Exception, so we
can't just catch a C++ exception and raise its swigged version.
Now the swigged C++ exception is wrapped in an OOFPyError, which
is derived from Exception, and that is raised instead.  It becomes
a little more difficult for python code to catch some kinds of
C++ exceptions and not others, and a bit ugly when Python needs
to throw an exception itself.  Perhaps we can have SWIG generate
objects that look like the C++ exceptions but are really just
OOFPyError instances.