The default preconditioner for all iterative solvers is ILUT.
The BiCG solver used to be just an alias for BiCGStab. (Eigen only
offers BiCGStab, and we kept the BiCG class for backward compatibility.)
To make the interface cleaner I made BiCG a secret registered class.
When called from scripts it should still invoke BiCGStab.

I accidentally included the new SRC/engine/matrixmethod.py in
the previous commit about to_string, and then I tried to have
git move it to this commit, but I don't think it worked.