Minor fixes.
The gfxwindow was trying to sort the toolboxes using the wrong key.

Added a TODO about adding a helpful message to an empty graphics window.

Fixed some integer divisions in oof.py.  In python2, 3/2=1, but
in python3, 3/2=1.5 and 3//2=1.  Finding the places where we
actually want integer division might be tricky.  In this case, it
was used to compute the size of intro banner in text mode.