Removed all StringVec mentions in .swg files.
When returning a std::vector<std::string>* to python, it's not
necessary to use a typedef for std::vector<std::string>.  Using
the StringVec typedef seemed to be hiding the typemap.  This had
been confusing me earlier.  Using the %newfree typemap and %newobject
does the correct thing when the std::vector being returned is a
new object.