New mechanism for translating and transmitting exceptions between Python and C++.
This is motivated by the need to have all Python exceptions be
derived from Exception or BaseException, which a swigged C++ exception
is not.

It's not thoroughly tested yet, but the first few relevant tests
in fundamental_test.py pass.  Exceptions defined outside of common
haven't been fully converted.

Instead of giving each C++ exception a pythonequiv() method which
was eval()ed to create a Python exception, there's now a
WRAP_OOFERROR_CLASS macro that defines a Python exception class for
a C++ ErrError subclass. The name of the wrapper class is the name
of the C++ class with Py prepended.  The wrapper class is derived
from both Exception and the swigged C++ class.  Being derived from
the C++ class allows it to be passed in to a swigged C++ function
and to come out the other side as a C++ exception.  Being derived
from Exception allows it to be raised and caught in Python code.

ErrError is now derived from PythonExportable, so all subclasses need
a classname() method.  This allows the %exception typemap to pass
the correct object out to Python.