Import path fixing.
For reasons I sort of understand, UTILS/file_utils.py was being
imported from regression.py and fundamental_test.py with different
module paths: one of them was storing it in sys.modules() as
TEST.UTILS.file_utils and the other as oof2.TEST.UTILS.file_utils.
This meant that the code in the module was being run twice,
and the referencedir value that was set after regression.py imported
it was being reset after fundamental_test.py imported it.

The solution is to use an absolute import.  This change will have to
be made in all test files, probably.