Small changes.
The library routine used to sanitize file names before using
markup in the file selector widget doesn't exist in python 3,
so I wrote a simple replacement.  I don't know if it's even necessary.

Added a TODO to menuparser.py.  Because input files might be binary,
they're opened with mode="rb", which means that strings read from
them are really bytes objects, and are ugly when printed.

Added a printAsString function in utils.py, and then didn't use it.
It might be a way of avoiding the above-mentioned TODO.