Backporting random code from the python3-conversion branch.
Don't use Python's random module for generating random numbers.
Use SRC/common/random.C instead, which is swigged as "crandom".

The random module in Python3 is different from Python2, and
tests which depend on random numbers are failing on Python3.
In order to check the Python3 version against the Python2 version,
they need to be using the same random numbers, so *both* programs
have to be changed.  The reference files will be updated in
Python2, and then compared to in Python3.

The code still uses Python's random.shuffle method, because
it can be told what random number generator to use.

This branch is not meant for release.