Added a shuffle function to crandom.spy, and changed all calls to random.shuffle.
It was sufficient to pass our own random number generator to
random.shuffle in order to get the same results in python2 and
python3, but the optional argument to random.shuffle that allows
that to work is deprecated and will be removed in python 3.11.
So I wrote a python version of the shuffle template in random.h,
put it in crandom.spy, and added a test to fundamental_test.py.

Now the python random module is not loaded by any OOF2 files,
except some old auxiliary scripts.

The regression tests don't pass, as expected, because all random
numbers that used to be generated by random.random() or random.randint()
are now generated by random.C.