Debugging CMakeLists.txt, which isn't working on macOS Monterey.
When linking swig-generated modules, it's looking for liboofcanvas
in the wrong location, even though the full path is being provided.

I'm committing now because the versions on the desktop and laptop
are getting confusing, so I'm consolidating the changes.  Also I
want to test the changes made to CMakeLists.txt on the desktop,
even though they don't work on the laptop.