Fixed problems with the MaterialType typemaps.
MaterialType needs to be converted from an Enum to a string when
calling C++ from Python.  I don't understand how it worked in
the python2/swig1.1 version.  Maybe it failed but since we
don't use its value we never noticed...  Anyhow the conversion wasn't
being done and was crashing with python3/swig4.

Also, the FETCH_CLASS swig macro was choking when trying to get
the MaterialType class for the "out" typemap, possibly because
the class path contained "SWIG" (ooflib.SWIG.engine.material).
Since MaterialType doesn't have to be in the same module as Material,
I put it in its own file.