Fixed a bunch of problems related to getting the Relaxation test to run.
The test still doesn't pass, but at least it runs.

Added new typemaps to typemaps.swg for swigging vectors of
PythonExportable objects.

Added a missing "%import node.swg" which was causing something in
the Element class to fail.

Fixed the monkeypatching of the swigged FEMesh constructor.
SWIG 4 requires this to be done differently.  Added a comment about
maybe doing it better.

Fixed the imports in material.spy, which should have been updated
when the property tree was renamed to "properties" and properties.swg
became property.swg.  Property.swg should probably be renamed again,
since "property" has a special meaning in python.

The Mesh destructor was looping over and removing boundary conditions
with "for bc in self.bdyconditions.values()", which doesn't work
anymore because dict.values() returns an iterator, not a list, and
can't be changed while iterating.  It has to make an explicit copy
with "for bc in list(self.bdyconditions.values())".

The following changes don't actually affect the Relaxation test
and probably should be in a separate commit, but they're not:

Added a comment to progress.C about what can cause the assert
statement in the Progress destructor to fail.  It can be quite
misleading.

Made the arguments to switchboard.verbose() optional, making it
easier to turn it on manually when debugging.

Cleaned up Element::getBndyEdge(), but didn't make any substantive
changes in the end.

Added a comment to property.swg to make it easier to find functions
whose names are constructed by swig macros and seem not to be defined
if you're just grepping for them.