Changed some sets and dicts to OrderedSets
In Python2, sets and dicts didn't preserve order -- looping over
their contents didn't address the items in a fixed order.  In
Python3, that order changed, so tests whose results depended on
the order failed.  To check the Python3 version against the Python2
the order must be the same, so I replaced some sets and dicts (used
as sets) with utils.OrderedDict in both versions.

After the conversion to Python3 is done we can change the OrderedSets
to dicts (used as sets) to dicts, because Python3 dicts preserve
the order of their items (as of 3.7).  But that will break the
tests again.