Fixes to skeleton boundary code.
The order in which the segments of closed boundaries were saved in
a data file depended on the order in which they were retrieved from
a Python set, and that changed in Python3, breaking some tests.
Now the order is deterministic.

The _sequenceable flag in boundary objects was an int but should
have been a bool.