Cleaned up ViscoElasticity.
It had been using its own typemap to convert the IsotropicRank4TensorCij
parameter to a Cijkl object.  Now it does it using a wrapper class
that calls the parameter's tensorForm method, like other Cijkl-using
Properties.

The C++ ViscoElasticityProp class has been renamed CViscoElasticity,
and the wrapper class is ViscoElasticity.  This is also in line with
the other Properties.