Fixed some swig problems wrt Materials and Properties.
Material.properties() wasn't returning properly swigged objects,
and monkeypatched Property methods weren't available for them.
But Material already has a python list of properties, so there's no
need for the properties() method.  I changed the name of the python
list from proplist to properties and deleted the method.

Also added a line to limit the scope of the %rename in property.swg.
Without the limitation, when property.swg was %imported or %included
into another swig file, all class methods in the importing file
were being renamed.