Various minor but nevertheless mostly necessary changes.
Pass True and False instead of 1 and 0 for boolean args to swigged functions.

Don't use CompoundFieldRefVec or MAKE_LISTVEC_TYPEMAPS(CompoundField).
Use the typemap in field.swg.

Deleted a questionable TODO and some 3D code.

Changed some *.next()s to *.increment()s that I had missed earlier
when fixing some of the not-really-a-python-iterator classes.