Updated the PyProperty classes, which are simpler with Swig4.
The swigged arguments passed to the python functions are now
constructed completely in C++, which means we don't need the wrapper
functions that used to be in pyproperty.spy.

Also, many functions had an unnecessary "referent" argument, which
was just a pointer to the PyObject containing the Property.  That
pointer had already been stored in C++ by the Property constructor.
Some methods also have a Property* argument, which I think can also
be removed.  I added a TODO about it.