Create boundaries in the same order as the python2 version did, so that the GUI tests don't object.
Skeleton boundaries are stored in a dictionary, and were being
returned in a different order in python 3 than they were in python
2 when they were being listed in the GUI.  This was making tests
fail.

Because the python3 dictionaries are ordered, the easiest fix was
to just create the boundaries in the constructor, before creating
the nodes and edges that comprise the boundaries.  That made the
various makeBoundary routines fail, because they were testing to
be sure that boundaries didn't already exist.  But that was the
wrong scope for those tests, so I just deleted them.  Unique boundary
naming is still enforced by the gui and menu commands.