Partial fixes to the gtklogger gui.
The gui is started using the version of python that is running
the parent oof2 process.

The filename passed to the gui can be either str, bytes, or PathLike
(the types accepted by the built-in open()).

Replaced some positional arguments by keywords arguments where
Gtk was issuing deprecation warnings.  I don't know why we're not
getting thousands more warnings, if positional arguments are really
deprecated.