Minor fixes.
Subproblem sorting is fixed.  The Python3 sorted() needs a key arg.

The bool arg in image.negate() needs to be a bool, not an int.
(Perhaps we should add a swig typemap so that bool args can be
passed as ints?)

Converted more C++ exceptions raised in Python to their Python
equivalents.