Did a TODO, added to another one, and made some output prettier.
Did the TODO about using os.stat.

Added to the TODO about different FileInput classes for ascii
and binary files, and changed it to a "TODO?".

The debug output from ascii data files was ugly, with a prepended
'b' because the characters hadn't been converted from bytes. Now
they're converted, and if the line is too long characters are
removed from the middle and the end is shown, after "...".