Two TODOs done in parameter.py
StringParameter checks that its value is not bytes.  I don't think
it can be, but it doesn't hurt to check.

RegisteredParameter checks that its "reg" constructor argument is a
class, not a instance.  Its regname() method had code for the case
in which reg was an instance, but it should never be one, so I
deleted that code, and added the constructor check.