Tutorials now use pango markup, instead of our hacked up BOLD() tag.
I'm not sure that the tutorials look any better, but now there's
more flexibility to improve them.   They need a review for consistency,
since the markup is used differently in different tutorials.

The tutorial source code is unfortunately less readable, because I took out all of the line breaks within paragraphs and they all wrap around.  But when adding text to a Gtk.TextBuffer, all whitespace in the buffer is displayed, so we can't have extra line breaks or indentation.  We used to fix that in the tutorial gui code while we were processing the BOLD tags, but I think it's better to let Gtk/Pango do all the work.  It will reduce future confusion.