Some gui test fixes.
The regex used to search for floats needs to be a raw string.  I
don't know why it worked in earlier versions of python.

The format for syntax errors is slightly different in python 3.9 than
in 3.8.  The caret ^ that points to the error is one space to the right.
I just changed the tests so that they only look at the part of the string
after the difference.
(generics.py had contained code that handled different syntax error
formats between python 2.5 and 2.6, but would have incorrectly made
a distinction between 3.5 and 3.6 as well, so it's a good thing
we skipped those versions.)