Modified the error message test to accomodate different python versions.
The format of some error messages depends on the python version,
and tests check for error messages have to be able to accept them
all.  The errorMsg() function in generics.py now takes multiple
strings, and the test passes if any of the strings are found in the
error dialog box.   This is not ideal, but is probably good enough.

The current test files work correctly for python 3.10, but messages
for 3.8 and 3.9 still have to be added.