Changed the errorMsg test so that it can accept multiple versions of a message.
The format of error messages is different in different versions of
Python.  The tests now pass for versions 3.8 through 3.10.  The
test strings are now in the tests.py file, where they can be multiline
strings, isntead of in the test.log file, which because it's not a
real python file required the strings to be one line, which made
them pretty much unreadable.  Also, there now comments indicating
which error message goes with which python version.