Added a check for defunct Skeletons when checking groups.
When a Skeleton is being destroyed, asynchronicity in the
switchboard callbacks can lead to parts of the GUI being
updated when the Skeleton is partially destroyed.  This was
causing intermittent errors in the GUI tests.  Adding the
check here appears to have eliminated the error, but I don't
have proof of that.

I think there's a general problem with switchboard signals.  It
maybe shouldn't be possible for a signal to induce another signal.
When a thread temporarily moves to a subthread, it allows some
other waiting thread to run on the main loop, and this can lead
to synchronization issues.